// src/backend/dto/create-events.dto.ts
export class CreateEventDto {
  branch_id: string
  end_date: string;
  end_time: string;
  exdTime_token: string;
  hour_event: string;
  level_event: string;
  exdDate_token: string;
  quota: number;
  start_date: string;
  start_time: string;
  term: string;;
  title: string;
  venue: string;
  year: string;
}